refactor: tests minor fixes
