+++ test/resources/discussion.ts
